fix bugs in incremental parse
* build/wisitoken-keys.el (wisitoken-goto-stack-trace): new
  (wisitoken-dtrt): use it

* wisitoken-lexer.adb (Full_Image): new

* wisitoken-parse-lr-mckenzie_recover-explore.adb (Try_Insert_Quote): fix bug, match changes
  (Try_Delete_Input): handle empty nonterm

* wisitoken-parse-lr-mckenzie_recover-parse.adb (Left_Breakdown): rename from Breakdown, copy fixes from Syntax_Trees.Left_Breakdown
  (Delete_Current_Applies): handle empty nonterm
  (Peek_Current_First_Shared_Terminal): add Following_Element parameter
  (Do_Delete): fix bug, match changes

* wisitoken-parse-lr-mckenzie_recover.adb (Peek_Shared_Start): match changes to To_Stream_Node_Parents
  (Delete_Check): match changes

* wisitoken-parse-lr-parser.adb (Finish_Parse): don't count non_grammar; Delete_Token that has non-grammar restores it.

* wisitoken-parse-lr-parser_lists.adb (Compare): fix bug

* wisitoken-parse-lr.ads (configuration): improve comment

* wisitoken-parse.adb (Next_Grammar_Token): use lexer.full_image
  (Edit_Tree): improve verbosity
  (Edit_Tree Do_Scan compute): simplify, fix bugs

* wisitoken-syntax_trees.adb (Next_Source_Terminal): new
  (Check_Non_Grammar): handle eoi
  (Delete_Subtree): alphabetize
  (Empty_Line): new
  (Find_Char_Pos): rename Include_Non_Grammar to Trailing_Non_Grammar
  (Find_New_Line): add post condition
  (Find_New_Line (out buffer_pos)): new
  (Left_Breakdown): fix bugs; delete empty nonterms
  (Line_At_Byte_Pos): new
  (Line_Begin_Token): handle empty lines
  (Line_Region_Internal): improve handling of SOI, handle empty nonterm ref
  (Line_Region): use to_stream_node_Parents to handle first_terminal ref
  (To_Stream_Node_Parents): handle rooted and first_terminal

+++ wisitoken-parse-lr-mckenzie_recover.adb
@@ -772,8 +772,10 @@ package body WisiToken.Parse.LR.McKenzie_Recover is
       return State : Peek_Shared_State (Config.Input_Stream'Access) do
          Parse.First_Shared_Terminal (Tree, State.Input_Terminal);
 
-         State.Shared_Terminal := Syntax_Trees.To_Stream_Node_Parents (Config.Current_Shared_Token);
-         Tree.First_Shared_Terminal (State.Shared_Terminal);
+         State.Shared_Terminal := Tree.To_Stream_Node_Parents (Config.Current_Shared_Token);
+         if Syntax_Trees.Rooted (State.Shared_Terminal.Ref) then
+            Tree.First_Shared_Terminal (State.Shared_Terminal);
+         end if;
       end return;
    end Peek_Shared_Start;
 
@@ -838,8 +840,12 @@ package body WisiToken.Parse.LR.McKenzie_Recover is
       Config : in out Configuration;
       ID     : in     Token_ID)
    is
-      Node : constant Syntax_Trees.Node_Access := Parse.Peek_Current_First_Shared_Terminal (Tree, Config);
+      Node : constant Syntax_Trees.Node_Access := Parse.Peek_Current_First_Shared_Terminal
+        (Tree, Config, Following_Element => False);
    begin
+      if Node = Syntax_Trees.Invalid_Node_Access then
+         raise Bad_Config;
+      end if;
       Delete_Check (Tree, Config, Node, ID);
    end Delete_Check;
